Update README.md
We are not responsible for any of the mods contained in this modpack. We are able to troubleshoot issues within the server and are only obliged to to fix problems with the server. If issues occur in your single player game, it is up to you to contact the developers of the mod/s to find the help you may require.